Changes message origin from product-service to product-scraper. Updates handler to fix test path reference and validates the event passed. Relocates some file and fixes need to reference aws-sdk. Renames LogGroup and RoleName to include service name by variable and starts these names with stage. Fixes handler result function to display results and errors. Renamed npm script from "test:scrape" to "sim:scrape".
