+++ products/lambda/package.json
@@ -5,7 +5,6 @@
   "author": "Nordstrom",
   "license": "Apache-2.0",
   "dependencies": {
-    "aws-sdk": "^2.7.16",
     "js-yaml": "^3.7.0"
   }
 }
